+++ CHANGELOG.md
@@ -33,6 +33,7 @@ calls to JPL Horizons.
 - Improved readability, functionality, and memory usage in `read_mwa_corr_fits`.
 
 ### Fixed
+- A bug in writing calfits files when the optional `time_range` parameter is not set.
 - A bug where the `TIME_CENTROID` field was not being filled in data sets written by
   `UVData.write_ms`, which caused odd behavior in some CASA routines (e.g., `gaincal`).
 - A bug in reading in uvfits files with baseline coordinates that have suffixes of